requirements.txt
asgiref==3.5.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
Django==4.0.4
django-allauth==0.50.0
django-channels==0.7.0
google-api-core==2.7.2
google-api-python-client==2.44.0
google-auth==2.6.5
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
httplib2==0.20.4
idna==3.3
mysql-client==0.0.1
mysql-connector-python==8.0.28
mysqlclient==2.1.0
oauthlib==3.2.0
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.8
python3-openid==3.2.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
